JDP200402-12 Encja produkt - lombok, uporządkowanie pakietów
