+++ src/main/java/com/kodilla/ecommercee/dao/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.dao;
 
+import com.kodilla.ecommercee.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
